Добавляет модуль валидации формы
